Moving to localhost.

+++ punk.py
@@ -49,8 +49,12 @@ if action.lower().strip() == 'save':
 		return False #because you finished the search without finding anything
 		
 	def sendChunk(tempfilename,name,tdl):
+		r = open(tempfilename,'rb')
+		piece = r.read()
 		csum = '|'+str(len(piece))+'|'+hashlib.md5(piece).hexdigest()+'|'+hashlib.sha1(piece).hexdigest()+'|'+hashlib.sha512(piece).hexdigest()
-		res = subprocess.check_output('curl -F "authorizationKey='+ad+'" -F "handler=1" -F "handlerNeeded=DataIntake" -F "uploadedfile=@'+tempfilename+'" http://futuramerlin.com/d/r/active.php', shell = True).strip()
+		#help from http://unix.stackexchange.com/questions/94604/does-curl-have-a-timeout
+		res = subprocess.check_output('curl --connect-timeout 30 -m 180 -F "authorizationKey='+ad+'" -F "handler=1" -F "handlerNeeded=DataIntake" -F "uploadedfile=@'+tempfilename+'" http://localhost:8888/d/r/active.php', shell = True).strip()
+		print res
 		if not re.match('[0-9]+\|',res.strip()):
 			sys.exit("Could not send data to server; please make a new snapshot later to continue.")
 		if res[res.find('|'):] != csum:
@@ -58,24 +62,27 @@ if action.lower().strip() == 'save':
 		return res
 	
 	def send(name,w,tdl,timedb):
-		#format: filename, res of metadata, res of each chunk
 		filenm = base64.b64encode(name)
+		print 'Working with new time database: '+timedb
 		os.system('grep '+filenm+' '+timedb+' > ./.filtered.punktimedb')
 		fb = open('./.filtered.punktimedb','rb')
 		timeopt = fb.read()
-		#based on http://stackoverflow.com/questions/11833266/how-do-i-read-the-first-line-of-a-string
+		print 'timeopt: '+timeopt
 		timemod = timeopt.split('\n',1)[0]
-		timemodb = timemod[timemod.find('|'):]
+		print 'timemod: '+timemod
+		timemodb = timemod[(timemod.find('|')+1):]
 		rtm = os.path.getmtime(name)
-		if rtm == timemodb:
+		print 'Last modified: '+str(rtm)
+		print 'Recorded modification time: '+timemodb
+		if str(rtm) == str(timemodb):
 			print 'Skipping unchanged file'
 			return
-		command = "sed -i.bak 's/^"+filenm.replace('.','\\.')+"*$//' "+timedb
-		wr = open(timedb,'wb')
-		wr.write(filenm+'|'+str(os.path.getmtime(name)))
-		command = "sed -i.bak 's/^"+filenm.replace('.','\\.')+"*$//' .snapshots.punkset/"+now+".punkdb"
+		command = "sed -i.punkbaktimedb 's/^"+filenm.replace('.','\\.')+"*$//' "+timedb
+		os.system(command)
+		command = "sed -i.punkbakdb 's/^"+filenm.replace('.','\\.')+"*$//' .snapshots.punkset/"+now+".punkdb"
+		os.system(command)
 		print name
-		if (name.startswith('./.snapshots.punkset/') and name.endswith('.punkdb')) or name == './.latest.punksr' or name == './.snapshots.punkset' or name == tempDir+'/.temp.punkd' or name == './.temp.punkp' or name == './.this.punkak':
+		if (name.startswith('./.snapshots.punkset/') and name.endswith('.punkbaktimedb')) or (name.startswith('./.snapshots.punkset/') and name.endswith('.punkbakdb')) or (name.startswith('./.snapshots.punkset/') and name.endswith('.punkdb')) or (name.startswith('./.snapshots.punkset/') and name.endswith('.punktimedb')) or name == './.latest.punksr' or name == './.temp.punksp' or name == './.filtered.punktimedb' or name == './.snapshots.punkset' or name == tempDir+'/.temp.punkd' or name == './.temp.punkp' or name == './.this.punkak':
 			print 'Skipping punk database file'
 			return
 		os.system('tar -c -f '+tempDir+'/.temp.punkd --no-recursion --format pax '+name)
@@ -84,19 +91,23 @@ if action.lower().strip() == 'save':
 		lenp = os.path.getsize(tempDir+'/.temp.punkd')
 		lenr = lenp - lenf
 		#based on http://www.unix.com/shell-programming-and-scripting/66466-remove-first-n-bytes-last-n-bytes-binary-file-aix.html
-		os.system('dd if='+tempDir+'/.temp.punkd of='+tempDir+'/.temp.punksp bs=1 count='+lenr)
+		os.system('dd if='+tempDir+'/.temp.punkd of='+tempDir+'/.temp.punksp bs=1 count='+str(lenr))
 		resmeta = sendChunk(tempDir+'/.temp.punksp',filenm,tdl)
 		resf = filenm+'|'+resmeta+'\n'
+		wr = open(timedb,'ab')
+		wr.write(filenm+'|'+str(os.path.getmtime(name)))
+		#help from http://stackoverflow.com/questions/3204782/how-to-check-if-a-file-is-a-directory-or-regular-file-in-python
 		w.write(resf)
-		f = open(name)
-		for piece in read_in_chunks(f):
-			wr = open('.temp.punkp', 'wb')
-			wr.write(piece)
-			wr.close()
-			resp = sendChunk('.temp.punkp')
-			resf = filenm+'|'+resp+'\n'
-			w.write(resf)
-		f.close()
+		if os.path.isfile(name):
+			f = open(name)
+			for piece in read_in_chunks(f):
+				wr = open('.temp.punkp', 'wb')
+				wr.write(piece)
+				wr.close()
+				resp = sendChunk('.temp.punkp',filenm,tdl)
+				resf = filenm+'|'+resp+'\n'
+				w.write(resf)
+			f.close()
 
 	# based on http://stackoverflow.com/questions/120656/directory-listing-in-python
 	now = strftime("%Y.%m.%d.%H.%M.%S.%f.%z", gmtime())
@@ -107,8 +118,12 @@ if action.lower().strip() == 'save':
 	else:
 		w = open('.latest.punksr', 'rb')
 		tdl = w.read()
-		os.system('cp .snapshots.punkset/'+tdl+'.punkdb .snapshots.punkset/'+now+'.punkdb')
-		os.system('cp .snapshots.punkset/'+tdl+'.punktimedb .snapshots.punkset/'+now+'.punktimedb')
+		os.system('cp -v .snapshots.punkset/'+tdl+'.punkdb .snapshots.punkset/'+now+'.punkdb')
+		print 'Working with existing time database: '+'.snapshots.punkset/'+tdl+'.punktimedb'
+		os.system('cp -v .snapshots.punkset/'+tdl+'.punktimedb .snapshots.punkset/'+now+'.punktimedb')
+	w = open('.latest.punksr', 'wb')
+	w.write(now)
+	w.close()
 	#help from http://stackoverflow.com/questions/273192/check-if-a-directory-exists-and-create-it-if-necessary
 	if not os.path.exists('.snapshots.punkset'):
 		os.makedirs('.snapshots.punkset')
@@ -126,9 +141,7 @@ if action.lower().strip() == 'save':
 			send(cfilename, w, tdl, timedb)
 	nres = "Completed snapshot at "+strftime("%Y.%m.%d.%H.%M.%S.%f.%z", gmtime())+"."
 	w.write(nres)
-	w = open('.latest.punksr', 'wb')
-	w.write(now)
-	w.close()
+
 	sys.exit(nres)
 
 if action.lower().strip() == 'restore':
